requirements.txt
Flask==2.1.1
requests==2.27.1
wheezy.template==3.1.0
Flask-SSE==1.0.0
APScheduler==3.9.1
rdflib==6.1.1
SPARQLWrapper==2.0.0